requirements.txt
streamlit
diffusers
transformers
torch
Pillow
requests